+++ apps/web/pages/api/registry/entries.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { db } from 'apps/web/src/utils/ocsRegistry';
+import { vercelDb as db } from 'apps/web/src/utils/datastores/rds';
 import { kv } from 'apps/web/src/utils/datastores/kv';
 import { logger } from 'apps/web/src/utils/logger';
 import { withTimeout } from 'apps/web/pages/api/decorators';
