+++ apps/web/src/utils/proofs/proofs_storage.ts
@@ -1,9 +1,5 @@
-import { createKysely } from '@vercel/postgres-kysely';
+import { vercelDb as db } from 'apps/web/src/utils/datastores/rds';
 import { Address } from 'viem';
-
-type Database = {
-  proofs: ProofsTable;
-};
 
 export enum ProofTableNamespace {
   UsernamesEarlyAccess = 'usernames_early_access',
@@ -12,14 +8,6 @@ export enum ProofTableNamespace {
   CBIDDiscount = 'basenames_cbid_discount',
 }
 
-type ProofsTable = {
-  address: Address;
-  namespace: string;
-  proofs: string;
-};
-
-//username_proofs
-
 const proofTableName = 'proofs';
 
 export async function getProofsByNamespaceAndAddress(
@@ -27,9 +15,7 @@ export async function getProofsByNamespaceAndAddress(
   namespace: ProofTableNamespace,
   caseInsensitive = true, // set false for big data sets
 ) {
-  let query = createKysely<Database>()
-    .selectFrom(proofTableName)
-    .where('namespace', '=', namespace.valueOf());
+  let query = db.selectFrom(proofTableName).where('namespace', '=', namespace.valueOf());
 
   /**
    * use = when possible to search by namespace_address index otherwise it's a text based search.
